CV_Analysis_Tool/requirements.txt
pdfminer.six==20201018
PyMuPDF==1.18.5
python-docx==0.8.10
pandas==1.1.3
nltk==3.5
spaCy==2.3.2
